[SECURITY-1158] Fold tests back into normal files
